@@ -1,6 +1,6 @@
-import { NgModule } from '@angular/core';
+import { ModuleWithProviders, NgModule } from '@angular/core';
 import { QUERY_CLIENT } from './angular-query.token';
-import { QueryCache, QueryClient } from './query/core';
+import { QueryClient } from './query/core';
 import { UseBaseQueryService } from './use-base-query/use-base-query.service';
 import { UseMutationService } from './use-mutation/use-mutation.service';
 import { UseQueryService } from './use-query/use-query.service';
@@ -9,7 +9,11 @@ import { UseQueryService } from './use-query/use-query.service';
   providers: [UseBaseQueryService, UseQueryService, UseMutationService],
 })
 export class AngularQueryModule {
-  static forRoot(queryCache = new QueryCache(), queryClient = new QueryClient({ queryCache })) {
+  static forRoot(
+    queryClient = new QueryClient()
+  ): ModuleWithProviders<AngularQueryModule> {
+    queryClient.mount();
+
     return {
       ngModule: AngularQueryModule,
       providers: [{ provide: QUERY_CLIENT, useValue: queryClient }],
