@@ -1,19 +1,19 @@
-export { QueryCache } from './queryCache'
-export { QueryClient } from './queryClient'
-export { QueryObserver } from './queryObserver'
-export { QueriesObserver } from './queriesObserver'
-export { InfiniteQueryObserver } from './infiniteQueryObserver'
-export { MutationCache } from './mutationCache'
-export { MutationObserver } from './mutationObserver'
-export { setLogger } from './logger'
-export { notifyManager } from './notifyManager'
-export { focusManager } from './focusManager'
-export { onlineManager } from './onlineManager'
-export { hashQueryKey, isError } from './utils'
-export { isCancelledError } from './retryer'
+export { QueryCache } from './queryCache';
+export { QueryClient } from './queryClient';
+export { QueryObserver } from './queryObserver';
+export { QueriesObserver } from './queriesObserver';
+export { InfiniteQueryObserver } from './infiniteQueryObserver';
+export { MutationCache } from './mutationCache';
+export { MutationObserver } from './mutationObserver';
+export { setLogger } from './logger';
+export { notifyManager } from './notifyManager';
+export { focusManager } from './focusManager';
+export { onlineManager } from './onlineManager';
+export { hashQueryKey, isError } from './utils';
+export { isCancelledError } from './retryer';
 
 // Types
-export * from './types'
-export type { CancelledError } from './retryer'
-export type { Query } from './query'
-export type { Logger } from './logger'
+export * from './types';
+export type { CancelledError } from './retryer';
+export type { Query } from './query';
+export type { Logger } from './logger';
