@@ -4,74 +4,74 @@ import type {
   InfiniteData,
   InfiniteQueryObserverOptions,
   InfiniteQueryObserverResult,
-} from './types'
-import type { QueryClient } from './queryClient'
-import { ObserverFetchOptions, QueryObserver } from './queryObserver'
+} from './types';
+import type { QueryClient } from './queryClient';
+import { ObserverFetchOptions, QueryObserver } from './queryObserver';
 import {
   hasNextPage,
   hasPreviousPage,
   infiniteQueryBehavior,
-} from './infiniteQueryBehavior'
+} from './infiniteQueryBehavior';
 
 type InfiniteQueryObserverListener<TData, TError> = (
   result: InfiniteQueryObserverResult<TData, TError>
-) => void
+) => void;
 
 export class InfiniteQueryObserver<
-  TData = unknown,
+  TQueryFnData = unknown,
   TError = unknown,
-  TQueryFnData = TData,
+  TData = TQueryFnData,
   TQueryData = TQueryFnData
 > extends QueryObserver<
-  InfiniteData<TData>,
-  TError,
   TQueryFnData,
+  TError,
+  InfiniteData<TData>,
   InfiniteData<TQueryData>
 > {
   // Type override
   subscribe!: (
     listener?: InfiniteQueryObserverListener<TData, TError>
-  ) => () => void
+  ) => () => void;
 
   // Type override
-  getCurrentResult!: () => InfiniteQueryObserverResult<TData, TError>
+  getCurrentResult!: () => InfiniteQueryObserverResult<TData, TError>;
 
   // Type override
   protected fetch!: (
     fetchOptions?: ObserverFetchOptions
-  ) => Promise<InfiniteQueryObserverResult<TData, TError>>
+  ) => Promise<InfiniteQueryObserverResult<TData, TError>>;
 
   // eslint-disable-next-line @typescript-eslint/no-useless-constructor
   constructor(
     client: QueryClient,
     options: InfiniteQueryObserverOptions<
-      TData,
-      TError,
       TQueryFnData,
+      TError,
+      TData,
       TQueryData
     >
   ) {
-    super(client, options)
+    super(client, options);
   }
 
   protected bindMethods(): void {
-    super.bindMethods()
-    this.fetchNextPage = this.fetchNextPage.bind(this)
-    this.fetchPreviousPage = this.fetchPreviousPage.bind(this)
+    super.bindMethods();
+    this.fetchNextPage = this.fetchNextPage.bind(this);
+    this.fetchPreviousPage = this.fetchPreviousPage.bind(this);
   }
 
   setOptions(
     options?: InfiniteQueryObserverOptions<
-      TData,
-      TError,
       TQueryFnData,
+      TError,
+      TData,
       TQueryData
     >
   ): void {
     super.setOptions({
       ...options,
-      behavior: infiniteQueryBehavior<TData, TError, TQueryFnData>(),
-    })
+      behavior: infiniteQueryBehavior<TQueryFnData, TError, TData>(),
+    });
   }
 
   fetchNextPage(
@@ -83,7 +83,7 @@ export class InfiniteQueryObserver<
       meta: {
         fetchMore: { direction: 'forward', pageParam: options?.pageParam },
       },
-    })
+    });
   }
 
   fetchPreviousPage(
@@ -95,25 +95,25 @@ export class InfiniteQueryObserver<
       meta: {
         fetchMore: { direction: 'backward', pageParam: options?.pageParam },
       },
-    })
+    });
   }
 
   protected getNewResult(
     willFetch?: boolean
   ): InfiniteQueryObserverResult<TData, TError> {
-    const { state } = this.getCurrentQuery()
-    const result = super.getNewResult(willFetch)
+    const { state } = this.getCurrentQuery();
+    const result = super.getNewResult(willFetch);
     return {
       ...result,
       fetchNextPage: this.fetchNextPage,
       fetchPreviousPage: this.fetchPreviousPage,
-      hasNextPage: hasNextPage(this.options, result.data?.pages),
-      hasPreviousPage: hasPreviousPage(this.options, result.data?.pages),
+      hasNextPage: hasNextPage(this.options, state.data?.pages),
+      hasPreviousPage: hasPreviousPage(this.options, state.data?.pages),
       isFetchingNextPage:
         state.isFetching && state.fetchMeta?.fetchMore?.direction === 'forward',
       isFetchingPreviousPage:
         state.isFetching &&
         state.fetchMeta?.fetchMore?.direction === 'backward',
-    }
+    };
   }
 }
