@@ -1,14 +1,14 @@
-import { noop } from './utils'
+import { noop } from './utils';
 
 // TYPES
 
 export interface Logger {
-  log: LogFunction
-  warn: LogFunction
-  error: LogFunction
+  log: LogFunction;
+  warn: LogFunction;
+  error: LogFunction;
 }
 
-type LogFunction = (...args: any[]) => void
+type LogFunction = (...args: any[]) => void;
 
 // FUNCTIONS
 
@@ -16,12 +16,12 @@ let logger: Logger = console || {
   error: noop,
   warn: noop,
   log: noop,
-}
+};
 
 export function getLogger(): Logger {
-  return logger
+  return logger;
 }
 
 export function setLogger(newLogger: Logger) {
-  logger = newLogger
+  logger = newLogger;
 }
