@@ -1,19 +1,19 @@
-import type { MutationOptions, MutationStatus } from './types'
-import type { MutationCache } from './mutationCache'
-import type { MutationObserver } from './mutationObserver'
-import { getLogger } from './logger'
-import { notifyManager } from './notifyManager'
-import { Retryer } from './retryer'
-import { noop } from './utils'
+import type { MutationOptions, MutationStatus } from './types';
+import type { MutationCache } from './mutationCache';
+import type { MutationObserver } from './mutationObserver';
+import { getLogger } from './logger';
+import { notifyManager } from './notifyManager';
+import { Retryer } from './retryer';
+import { noop } from './utils';
 
 // TYPES
 
 interface MutationConfig<TData, TError, TVariables, TContext> {
-  mutationId: number
-  mutationCache: MutationCache
-  options: MutationOptions<TData, TError, TVariables, TContext>
-  defaultOptions?: MutationOptions<TData, TError, TVariables, TContext>
-  state?: MutationState<TData, TError, TVariables, TContext>
+  mutationId: number;
+  mutationCache: MutationCache;
+  options: MutationOptions<TData, TError, TVariables, TContext>;
+  defaultOptions?: MutationOptions<TData, TError, TVariables, TContext>;
+  state?: MutationState<TData, TError, TVariables, TContext>;
 }
 
 export interface MutationState<
@@ -22,56 +22,56 @@ export interface MutationState<
   TVariables = void,
   TContext = unknown
 > {
-  context: TContext | undefined
-  data: TData | undefined
-  error: TError | null
-  failureCount: number
-  isPaused: boolean
-  status: MutationStatus
-  variables: TVariables | undefined
+  context: TContext | undefined;
+  data: TData | undefined;
+  error: TError | null;
+  failureCount: number;
+  isPaused: boolean;
+  status: MutationStatus;
+  variables: TVariables | undefined;
 }
 
 interface FailedAction {
-  type: 'failed'
+  type: 'failed';
 }
 
 interface LoadingAction<TVariables, TContext> {
-  type: 'loading'
-  variables?: TVariables
-  context?: TContext
+  type: 'loading';
+  variables?: TVariables;
+  context?: TContext;
 }
 
 interface SuccessAction<TData> {
-  type: 'success'
-  data: TData
+  type: 'success';
+  data: TData;
 }
 
 interface ErrorAction<TError> {
-  type: 'error'
-  error: TError
+  type: 'error';
+  error: TError;
 }
 
 interface PauseAction {
-  type: 'pause'
+  type: 'pause';
 }
 
 interface ContinueAction {
-  type: 'continue'
+  type: 'continue';
 }
 
 interface SetStateAction<TData, TError, TVariables, TContext> {
-  type: 'setState'
-  state: MutationState<TData, TError, TVariables, TContext>
+  type: 'setState';
+  state: MutationState<TData, TError, TVariables, TContext>;
 }
 
-type Action<TData, TError, TVariables, TContext> =
+export type Action<TData, TError, TVariables, TContext> =
   | ContinueAction
   | ErrorAction<TError>
   | FailedAction
   | LoadingAction<TVariables, TContext>
   | PauseAction
   | SetStateAction<TData, TError, TVariables, TContext>
-  | SuccessAction<TData>
+  | SuccessAction<TData>;
 
 // CLASS
 
@@ -81,87 +81,87 @@ export class Mutation<
   TVariables = void,
   TContext = unknown
 > {
-  state: MutationState<TData, TError, TVariables, TContext>
-  options: MutationOptions<TData, TError, TVariables, TContext>
-  mutationId: number
+  state: MutationState<TData, TError, TVariables, TContext>;
+  options: MutationOptions<TData, TError, TVariables, TContext>;
+  mutationId: number;
 
-  private observers: MutationObserver<TData, TError, TVariables, TContext>[]
-  private mutationCache: MutationCache
-  private retryer?: Retryer<TData, TError>
+  private observers: MutationObserver<TData, TError, TVariables, TContext>[];
+  private mutationCache: MutationCache;
+  private retryer?: Retryer<TData, TError>;
 
   constructor(config: MutationConfig<TData, TError, TVariables, TContext>) {
     this.options = {
       ...config.defaultOptions,
       ...config.options,
-    }
-    this.mutationId = config.mutationId
-    this.mutationCache = config.mutationCache
-    this.observers = []
-    this.state = config.state || getDefaultState()
+    };
+    this.mutationId = config.mutationId;
+    this.mutationCache = config.mutationCache;
+    this.observers = [];
+    this.state = config.state || getDefaultState();
   }
 
   setState(state: MutationState<TData, TError, TVariables, TContext>): void {
-    this.dispatch({ type: 'setState', state })
+    this.dispatch({ type: 'setState', state });
   }
 
   addObserver(observer: MutationObserver<any, any, any, any>): void {
     if (this.observers.indexOf(observer) === -1) {
-      this.observers.push(observer)
+      this.observers.push(observer);
     }
   }
 
   removeObserver(observer: MutationObserver<any, any, any, any>): void {
-    this.observers = this.observers.filter(x => x !== observer)
+    this.observers = this.observers.filter((x) => x !== observer);
   }
 
   cancel(): Promise<void> {
     if (this.retryer) {
-      this.retryer.cancel()
-      return this.retryer.promise.then(noop).catch(noop)
+      this.retryer.cancel();
+      return this.retryer.promise.then(noop).catch(noop);
     }
-    return Promise.resolve()
+    return Promise.resolve();
   }
 
   continue(): Promise<TData> {
     if (this.retryer) {
-      this.retryer.continue()
-      return this.retryer.promise
+      this.retryer.continue();
+      return this.retryer.promise;
     }
-    return this.execute()
+    return this.execute();
   }
 
   execute(): Promise<TData> {
-    let data: TData
+    let data: TData;
 
-    const restored = this.state.status === 'loading'
+    const restored = this.state.status === 'loading';
 
-    let promise = Promise.resolve()
+    let promise = Promise.resolve();
 
     if (!restored) {
-      this.dispatch({ type: 'loading', variables: this.options.variables! })
+      this.dispatch({ type: 'loading', variables: this.options.variables! });
       promise = promise
         .then(() => this.options.onMutate?.(this.state.variables!))
-        .then(context => {
+        .then((context) => {
           if (context !== this.state.context) {
             this.dispatch({
               type: 'loading',
               context,
               variables: this.state.variables,
-            })
+            });
           }
-        })
+        });
     }
 
     return promise
       .then(() => this.executeMutation())
-      .then(result => {
-        data = result
+      .then((result) => {
+        data = result;
       })
       .then(() =>
         this.options.onSuccess?.(
           data,
           this.state.variables!,
-          this.state.context
+          this.state.context!
         )
       )
       .then(() =>
@@ -173,11 +173,23 @@ export class Mutation<
         )
       )
       .then(() => {
-        this.dispatch({ type: 'success', data })
-        return data
+        this.dispatch({ type: 'success', data });
+        return data;
       })
-      .catch(error => {
-        getLogger().error(error)
+      .catch((error) => {
+        // Notify cache callback
+        if (this.mutationCache.config.onError) {
+          this.mutationCache.config.onError(
+            error,
+            this.state.variables,
+            this.state.context,
+            this as Mutation<unknown, unknown, unknown, unknown>
+          );
+        }
+
+        // Log error
+        getLogger().error(error);
+
         return Promise.resolve()
           .then(() =>
             this.options.onError?.(
@@ -195,45 +207,45 @@ export class Mutation<
             )
           )
           .then(() => {
-            this.dispatch({ type: 'error', error })
-            throw error
-          })
-      })
+            this.dispatch({ type: 'error', error });
+            throw error;
+          });
+      });
   }
 
   private executeMutation(): Promise<TData> {
     this.retryer = new Retryer({
       fn: () => {
         if (!this.options.mutationFn) {
-          return Promise.reject('No mutationFn found')
+          return Promise.reject('No mutationFn found');
         }
-        return this.options.mutationFn(this.state.variables!)
+        return this.options.mutationFn(this.state.variables!);
       },
       onFail: () => {
-        this.dispatch({ type: 'failed' })
+        this.dispatch({ type: 'failed' });
       },
       onPause: () => {
-        this.dispatch({ type: 'pause' })
+        this.dispatch({ type: 'pause' });
       },
       onContinue: () => {
-        this.dispatch({ type: 'continue' })
+        this.dispatch({ type: 'continue' });
       },
       retry: this.options.retry ?? 0,
       retryDelay: this.options.retryDelay,
-    })
+    });
 
-    return this.retryer.promise
+    return this.retryer.promise;
   }
 
   private dispatch(action: Action<TData, TError, TVariables, TContext>): void {
-    this.state = reducer(this.state, action)
+    this.state = reducer(this.state, action);
 
     notifyManager.batch(() => {
-      this.observers.forEach(observer => {
-        observer.onMutationUpdate()
-      })
-      this.mutationCache.notify(this)
-    })
+      this.observers.forEach((observer) => {
+        observer.onMutationUpdate(action);
+      });
+      this.mutationCache.notify(this);
+    });
   }
 }
 
@@ -251,7 +263,7 @@ export function getDefaultState<
     isPaused: false,
     status: 'idle',
     variables: undefined,
-  }
+  };
 }
 
 function reducer<TData, TError, TVariables, TContext>(
@@ -263,17 +275,17 @@ function reducer<TData, TError, TVariables, TContext>(
       return {
         ...state,
         failureCount: state.failureCount + 1,
-      }
+      };
     case 'pause':
       return {
         ...state,
         isPaused: true,
-      }
+      };
     case 'continue':
       return {
         ...state,
         isPaused: false,
-      }
+      };
     case 'loading':
       return {
         ...state,
@@ -283,7 +295,7 @@ function reducer<TData, TError, TVariables, TContext>(
         isPaused: false,
         status: 'loading',
         variables: action.variables,
-      }
+      };
     case 'success':
       return {
         ...state,
@@ -291,7 +303,7 @@ function reducer<TData, TError, TVariables, TContext>(
         error: null,
         status: 'success',
         isPaused: false,
-      }
+      };
     case 'error':
       return {
         ...state,
@@ -300,13 +312,13 @@ function reducer<TData, TError, TVariables, TContext>(
         failureCount: state.failureCount + 1,
         isPaused: false,
         status: 'error',
-      }
+      };
     case 'setState':
       return {
         ...state,
         ...action.state,
-      }
+      };
     default:
-      return state
+      return state;
   }
 }
