@@ -1,51 +1,51 @@
-import { scheduleMicrotask } from './utils'
+import { scheduleMicrotask } from './utils';
 
 // TYPES
 
-type NotifyCallback = () => void
+type NotifyCallback = () => void;
 
-type NotifyFunction = (callback: () => void) => void
+type NotifyFunction = (callback: () => void) => void;
 
-type BatchNotifyFunction = (callback: () => void) => void
+type BatchNotifyFunction = (callback: () => void) => void;
 
 // CLASS
 
 class NotifyManager {
-  private queue: NotifyCallback[]
-  private transactions: number
-  private notifyFn: NotifyFunction
-  private batchNotifyFn: BatchNotifyFunction
+  private queue: NotifyCallback[];
+  private transactions: number;
+  private notifyFn: NotifyFunction;
+  private batchNotifyFn: BatchNotifyFunction;
 
   constructor() {
-    this.queue = []
-    this.transactions = 0
+    this.queue = [];
+    this.transactions = 0;
 
     this.notifyFn = (callback: () => void) => {
-      callback()
-    }
+      callback();
+    };
 
     this.batchNotifyFn = (callback: () => void) => {
-      callback()
-    }
+      callback();
+    };
   }
 
   batch<T>(callback: () => T): T {
-    this.transactions++
-    const result = callback()
-    this.transactions--
+    this.transactions++;
+    const result = callback();
+    this.transactions--;
     if (!this.transactions) {
-      this.flush()
+      this.flush();
     }
-    return result
+    return result;
   }
 
   schedule(callback: NotifyCallback): void {
     if (this.transactions) {
-      this.queue.push(callback)
+      this.queue.push(callback);
     } else {
       scheduleMicrotask(() => {
-        this.notifyFn(callback)
-      })
+        this.notifyFn(callback);
+      });
     }
   }
 
@@ -55,22 +55,22 @@ class NotifyManager {
   batchCalls<T extends Function>(callback: T): T {
     return ((...args: any[]) => {
       this.schedule(() => {
-        callback(...args)
-      })
-    }) as any
+        callback(...args);
+      });
+    }) as any;
   }
 
   flush(): void {
-    const queue = this.queue
-    this.queue = []
+    const queue = this.queue;
+    this.queue = [];
     if (queue.length) {
       scheduleMicrotask(() => {
         this.batchNotifyFn(() => {
-          queue.forEach(callback => {
-            this.notifyFn(callback)
-          })
-        })
-      })
+          queue.forEach((callback) => {
+            this.notifyFn(callback);
+          });
+        });
+      });
     }
   }
 
@@ -79,7 +79,7 @@ class NotifyManager {
    * This can be used to for example wrap notifications with `React.act` while running tests.
    */
   setNotifyFunction(fn: NotifyFunction) {
-    this.notifyFn = fn
+    this.notifyFn = fn;
   }
 
   /**
@@ -87,10 +87,10 @@ class NotifyManager {
    * By default React Query will use the batch function provided by ReactDOM or React Native.
    */
   setBatchNotifyFunction(fn: BatchNotifyFunction) {
-    this.batchNotifyFn = fn
+    this.batchNotifyFn = fn;
   }
 }
 
 // SINGLETON
 
-export const notifyManager = new NotifyManager()
+export const notifyManager = new NotifyManager();
