@@ -1,47 +1,66 @@
-import { Subscribable } from './subscribable'
-import { isServer } from './utils'
+import { Subscribable } from './subscribable';
+import { isServer } from './utils';
 
 class OnlineManager extends Subscribable {
-  private removeHandler?: () => void
+  private online?: boolean;
+  private removeEventListener?: () => void;
 
   protected onSubscribe(): void {
-    if (!this.removeHandler) {
-      this.setDefaultHandler()
+    if (!this.removeEventListener) {
+      this.setDefaultEventListener();
     }
   }
 
-  setHandler(init: (onOnline: () => void) => () => void): void {
-    if (this.removeHandler) {
-      this.removeHandler()
+  setEventListener(
+    setup: (setOnline: () => void) => (online?: boolean) => void
+  ): void {
+    if (this.removeEventListener) {
+      this.removeEventListener();
     }
-    this.removeHandler = init(() => {
-      this.onOnline()
-    })
+    this.removeEventListener = setup((online?: boolean) => {
+      if (typeof online === 'boolean') {
+        this.setOnline(online);
+      } else {
+        this.onOnline();
+      }
+    });
+  }
+
+  setOnline(online?: boolean): void {
+    this.online = online;
+
+    if (online) {
+      this.onOnline();
+    }
   }
 
   onOnline(): void {
-    this.listeners.forEach(listener => {
-      listener()
-    })
+    this.listeners.forEach((listener) => {
+      listener();
+    });
   }
 
   isOnline(): boolean {
-    return navigator.onLine === undefined || navigator.onLine
+    if (typeof this.online === 'boolean') {
+      return this.online;
+    }
+
+    return navigator.onLine === undefined || navigator.onLine;
   }
 
-  private setDefaultHandler() {
+  private setDefaultEventListener() {
     if (!isServer && window?.addEventListener) {
-      this.setHandler(onOnline => {
+      this.setEventListener((onOnline) => {
         // Listen to online
-        window.addEventListener('online', onOnline, false)
+        window.addEventListener('online', onOnline, false);
 
         return () => {
           // Be sure to unsubscribe if a new handler is set
-          window.removeEventListener('online', onOnline)
-        }
-      })
+          window.removeEventListener('online', onOnline);
+        };
+      });
     }
   }
 }
 
-export const onlineManager = new OnlineManager()
+export const onlineManager = new OnlineManager();
