@@ -6,106 +6,106 @@ import {
   noop,
   replaceEqualDeep,
   timeUntilStale,
-} from './utils'
+} from './utils';
 import type {
   InitialDataFunction,
-  QueryFunction,
   QueryKey,
   QueryOptions,
   QueryStatus,
-} from './types'
-import type { QueryCache } from './queryCache'
-import type { QueryObserver } from './queryObserver'
-import { notifyManager } from './notifyManager'
-import { getLogger } from './logger'
-import { Retryer, CancelOptions, isCancelledError } from './retryer'
+  QueryFunctionContext,
+} from './types';
+import type { QueryCache } from './queryCache';
+import type { QueryObserver } from './queryObserver';
+import { notifyManager } from './notifyManager';
+import { getLogger } from './logger';
+import { Retryer, CancelOptions, isCancelledError } from './retryer';
 
 // TYPES
 
-interface QueryConfig<TData, TError, TQueryFnData> {
-  cache: QueryCache
-  queryKey: QueryKey
-  queryHash: string
-  options?: QueryOptions<TData, TError, TQueryFnData>
-  defaultOptions?: QueryOptions<TData, TError, TQueryFnData>
-  state?: QueryState<TData, TError>
+interface QueryConfig<TQueryFnData, TError, TData> {
+  cache: QueryCache;
+  queryKey: QueryKey;
+  queryHash: string;
+  options?: QueryOptions<TQueryFnData, TError, TData>;
+  defaultOptions?: QueryOptions<TQueryFnData, TError, TData>;
+  state?: QueryState<TData, TError>;
 }
 
 export interface QueryState<TData = unknown, TError = unknown> {
-  data: TData | undefined
-  dataUpdateCount: number
-  dataUpdatedAt: number
-  error: TError | null
-  errorUpdateCount: number
-  errorUpdatedAt: number
-  fetchFailureCount: number
-  fetchMeta: any
-  isFetching: boolean
-  isInvalidated: boolean
-  isPaused: boolean
-  status: QueryStatus
+  data: TData | undefined;
+  dataUpdateCount: number;
+  dataUpdatedAt: number;
+  error: TError | null;
+  errorUpdateCount: number;
+  errorUpdatedAt: number;
+  fetchFailureCount: number;
+  fetchMeta: any;
+  isFetching: boolean;
+  isInvalidated: boolean;
+  isPaused: boolean;
+  status: QueryStatus;
 }
 
-export interface FetchContext<TData, TError, TQueryFnData> {
-  state: QueryState<TData, TError>
-  options: QueryOptions<TData, TError, TQueryFnData>
-  params: unknown[]
-  fetchOptions?: FetchOptions
-  queryFn: QueryFunction
+export interface FetchContext<TQueryFnData, TError, TData> {
+  fetchFn: () => unknown | Promise<unknown>;
+  fetchOptions?: FetchOptions;
+  options: QueryOptions<TQueryFnData, TError, TData>;
+  queryKey: QueryKey;
+  state: QueryState<TData, TError>;
 }
 
 export interface QueryBehavior<
-  TData = unknown,
+  TQueryFnData = unknown,
   TError = unknown,
-  TQueryFnData = TData
+  TData = TQueryFnData
 > {
-  onFetch: (context: FetchContext<TData, TError, TQueryFnData>) => void
+  onFetch: (context: FetchContext<TQueryFnData, TError, TData>) => void;
 }
 
 export interface FetchOptions {
-  cancelRefetch?: boolean
-  meta?: any
+  cancelRefetch?: boolean;
+  meta?: any;
 }
 
 export interface SetDataOptions {
-  updatedAt?: number
+  updatedAt?: number;
 }
 
 interface FailedAction {
-  type: 'failed'
+  type: 'failed';
 }
 
 interface FetchAction {
-  type: 'fetch'
-  meta?: any
+  type: 'fetch';
+  meta?: any;
 }
 
 interface SuccessAction<TData> {
-  data: TData | undefined
-  type: 'success'
-  updatedAt?: number
+  data: TData | undefined;
+  type: 'success';
+  dataUpdatedAt?: number;
 }
 
 interface ErrorAction<TError> {
-  type: 'error'
-  error: TError
+  type: 'error';
+  error: TError;
 }
 
 interface InvalidateAction {
-  type: 'invalidate'
+  type: 'invalidate';
 }
 
 interface PauseAction {
-  type: 'pause'
+  type: 'pause';
 }
 
 interface ContinueAction {
-  type: 'continue'
+  type: 'continue';
 }
 
 interface SetStateAction<TData, TError> {
-  type: 'setState'
-  state: QueryState<TData, TError>
+  type: 'setState';
+  state: QueryState<TData, TError>;
 }
 
 export type Action<TData, TError> =
@@ -116,126 +116,141 @@ export type Action<TData, TError> =
   | InvalidateAction
   | PauseAction
   | SetStateAction<TData, TError>
-  | SuccessAction<TData>
+  | SuccessAction<TData>;
 
 // CLASS
 
-export class Query<TData = unknown, TError = unknown, TQueryFnData = TData> {
-  queryKey: QueryKey
-  queryHash: string
-  options!: QueryOptions<TData, TError, TQueryFnData>
-  state: QueryState<TData, TError>
-  cacheTime!: number
+export class Query<
+  TQueryFnData = unknown,
+  TError = unknown,
+  TData = TQueryFnData
+> {
+  queryKey: QueryKey;
+  queryHash: string;
+  options!: QueryOptions<TQueryFnData, TError, TData>;
+  initialState: QueryState<TData, TError>;
+  state: QueryState<TData, TError>;
+  cacheTime!: number;
 
-  private cache: QueryCache
-  private promise?: Promise<TData>
-  private gcTimeout?: number
-  private retryer?: Retryer<unknown, TError>
-  private observers: QueryObserver<any, any, any, any>[]
-  private defaultOptions?: QueryOptions<TData, TError, TQueryFnData>
+  private cache: QueryCache;
+  private promise?: Promise<TData>;
+  private gcTimeout?: number;
+  private retryer?: Retryer<unknown, TError>;
+  private observers: QueryObserver<any, any, any, any>[];
+  private defaultOptions?: QueryOptions<TQueryFnData, TError, TData>;
 
-  constructor(config: QueryConfig<TData, TError, TQueryFnData>) {
-    this.defaultOptions = config.defaultOptions
-    this.setOptions(config.options)
-    this.observers = []
-    this.cache = config.cache
-    this.queryKey = config.queryKey
-    this.queryHash = config.queryHash
-    this.state = config.state || this.getDefaultState(this.options)
-    this.scheduleGc()
+  constructor(config: QueryConfig<TQueryFnData, TError, TData>) {
+    this.defaultOptions = config.defaultOptions;
+    this.setOptions(config.options);
+    this.observers = [];
+    this.cache = config.cache;
+    this.queryKey = config.queryKey;
+    this.queryHash = config.queryHash;
+    this.initialState = config.state || this.getDefaultState(this.options);
+    this.state = this.initialState;
+    this.scheduleGc();
   }
 
   private setOptions(
-    options?: QueryOptions<TData, TError, TQueryFnData>
+    options?: QueryOptions<TQueryFnData, TError, TData>
   ): void {
-    this.options = { ...this.defaultOptions, ...options }
+    this.options = { ...this.defaultOptions, ...options };
 
     // Default to 5 minutes if not cache time is set
     this.cacheTime = Math.max(
       this.cacheTime || 0,
       this.options.cacheTime ?? 5 * 60 * 1000
-    )
+    );
   }
 
-  setDefaultOptions(options: QueryOptions<TData, TError, TQueryFnData>): void {
-    this.defaultOptions = options
+  setDefaultOptions(options: QueryOptions<TQueryFnData, TError, TData>): void {
+    this.defaultOptions = options;
   }
 
   private scheduleGc(): void {
-    this.clearGcTimeout()
+    this.clearGcTimeout();
 
     if (isValidTimeout(this.cacheTime)) {
       this.gcTimeout = setTimeout(() => {
-        if (!this.observers.length) {
-          this.cache.remove(this)
-        }
-      }, this.cacheTime)
+        this.optionalRemove();
+      }, this.cacheTime);
     }
   }
 
   private clearGcTimeout() {
-    clearTimeout(this.gcTimeout)
-    this.gcTimeout = undefined
+    clearTimeout(this.gcTimeout);
+    this.gcTimeout = undefined;
+  }
+
+  private optionalRemove() {
+    if (!this.observers.length && !this.state.isFetching) {
+      this.cache.remove(this);
+    }
   }
 
   setData(
     updater: Updater<TData | undefined, TData>,
     options?: SetDataOptions
   ): TData {
-    const prevData = this.state.data
+    const prevData = this.state.data;
 
     // Get the new data
-    let data = functionalUpdate(updater, prevData)
-
-    // Structurally share data between prev and new data if needed
-    if (this.options.structuralSharing !== false) {
-      data = replaceEqualDeep(prevData, data)
-    }
+    let data = functionalUpdate(updater, prevData);
 
     // Use prev data if an isDataEqual function is defined and returns `true`
     if (this.options.isDataEqual?.(prevData, data)) {
-      data = prevData as TData
+      data = prevData as TData;
+    } else if (this.options.structuralSharing !== false) {
+      // Structurally share data between prev and new data if needed
+      data = replaceEqualDeep(prevData, data);
     }
 
     // Set data and mark it as cached
     this.dispatch({
       data,
       type: 'success',
-      updatedAt: options?.updatedAt,
-    })
+      dataUpdatedAt: options?.updatedAt,
+    });
 
-    return data
+    return data;
   }
 
   setState(state: QueryState<TData, TError>): void {
-    this.dispatch({ type: 'setState', state })
+    this.dispatch({ type: 'setState', state });
   }
 
   cancel(options?: CancelOptions): Promise<void> {
-    const promise = this.promise
-    this.retryer?.cancel(options)
-    return promise ? promise.then(noop).catch(noop) : Promise.resolve()
+    const promise = this.promise;
+    this.retryer?.cancel(options);
+    return promise ? promise.then(noop).catch(noop) : Promise.resolve();
   }
 
   destroy(): void {
-    this.clearGcTimeout()
-    this.cancel()
+    this.clearGcTimeout();
+    this.cancel();
+  }
+
+  reset(): void {
+    this.destroy();
+    this.setState(this.initialState);
   }
 
   isActive(): boolean {
-    return this.observers.some(observer => observer.options.enabled !== false)
+    return this.observers.some(
+      (observer) => observer.options.enabled !== false
+    );
   }
 
   isFetching(): boolean {
-    return this.state.isFetching
+    return this.state.isFetching;
   }
 
   isStale(): boolean {
     return (
       this.state.isInvalidated ||
       !this.state.dataUpdatedAt ||
-      this.observers.some(observer => observer.getCurrentResult().isStale)
-    )
+      this.observers.some((observer) => observer.getCurrentResult().isStale)
+    );
   }
 
   isStaleByTime(staleTime = 0): boolean {
@@ -243,124 +258,125 @@ export class Query<TData = unknown, TError = unknown, TQueryFnData = TData> {
       this.state.isInvalidated ||
       !this.state.dataUpdatedAt ||
       !timeUntilStale(this.state.dataUpdatedAt, staleTime)
-    )
+    );
   }
 
   onFocus(): void {
-    const observer = this.observers.find(x => x.willFetchOnWindowFocus())
+    const observer = this.observers.find((x) => x.willFetchOnWindowFocus());
 
     if (observer) {
-      observer.refetch()
+      observer.refetch();
     }
 
     // Continue fetch if currently paused
-    this.retryer?.continue()
+    this.retryer?.continue();
   }
 
   onOnline(): void {
-    const observer = this.observers.find(x => x.willFetchOnReconnect())
+    const observer = this.observers.find((x) => x.willFetchOnReconnect());
 
     if (observer) {
-      observer.refetch()
+      observer.refetch();
     }
 
     // Continue fetch if currently paused
-    this.retryer?.continue()
+    this.retryer?.continue();
   }
 
   addObserver(observer: QueryObserver<any, any, any, any>): void {
     if (this.observers.indexOf(observer) === -1) {
-      this.observers.push(observer)
+      this.observers.push(observer);
 
       // Stop the query from being garbage collected
-      this.clearGcTimeout()
+      this.clearGcTimeout();
 
-      this.cache.notify(this)
+      this.cache.notify(this);
     }
   }
 
   removeObserver(observer: QueryObserver<any, any, any, any>): void {
     if (this.observers.indexOf(observer) !== -1) {
-      this.observers = this.observers.filter(x => x !== observer)
+      this.observers = this.observers.filter((x) => x !== observer);
 
       if (!this.observers.length) {
         // If the transport layer does not support cancellation
         // we'll let the query continue so the result can be cached
         if (this.retryer) {
           if (this.retryer.isTransportCancelable) {
-            this.retryer.cancel()
+            this.retryer.cancel();
           } else {
-            this.retryer.cancelRetry()
+            this.retryer.cancelRetry();
           }
         }
 
         if (this.cacheTime) {
-          this.scheduleGc()
+          this.scheduleGc();
         } else {
-          this.cache.remove(this)
+          this.cache.remove(this);
         }
       }
 
-      this.cache.notify(this)
+      this.cache.notify(this);
     }
   }
 
   invalidate(): void {
     if (!this.state.isInvalidated) {
-      this.dispatch({ type: 'invalidate' })
+      this.dispatch({ type: 'invalidate' });
     }
   }
 
   fetch(
-    options?: QueryOptions<TData, TError, TQueryFnData>,
+    options?: QueryOptions<TQueryFnData, TError, TData>,
     fetchOptions?: FetchOptions
   ): Promise<TData> {
     if (this.state.isFetching)
       if (this.state.dataUpdatedAt && fetchOptions?.cancelRefetch) {
         // Silently cancel current fetch if the user wants to cancel refetches
-        this.cancel({ silent: true })
+        this.cancel({ silent: true });
       } else if (this.promise) {
         // Return current promise if we are already fetching
-        return this.promise
+        return this.promise;
       }
 
     // Update config if passed, otherwise the config from the last execution is used
     if (options) {
-      this.setOptions(options)
+      this.setOptions(options);
     }
 
     // Use the options from the first observer with a query function if no function is found.
     // This can happen when the query is hydrated or created with setQueryData.
     if (!this.options.queryFn) {
-      const observer = this.observers.find(x => x.options.queryFn)
+      const observer = this.observers.find((x) => x.options.queryFn);
       if (observer) {
-        this.setOptions(observer.options)
+        this.setOptions(observer.options);
       }
     }
 
-    // Get the query function params
-    let params = ensureArray(this.queryKey)
+    // Create query function context
+    const queryKey = ensureArray(this.queryKey);
+    const queryFnContext: QueryFunctionContext = {
+      queryKey,
+      pageParam: undefined,
+    };
 
-    const filter = this.options.queryFnParamsFilter
-    params = filter ? filter(params) : params
-
-    // Get query function
-    const queryFn = () =>
+    // Create fetch function
+    const fetchFn = () =>
       this.options.queryFn
-        ? this.options.queryFn(...params)
-        : Promise.reject('Missing queryFn')
+        ? this.options.queryFn(queryFnContext)
+        : Promise.reject('Missing queryFn');
 
     // Trigger behavior hook
-    const context: FetchContext<TData, TError, TQueryFnData> = {
+    const context: FetchContext<TQueryFnData, TError, TData> = {
       fetchOptions,
       options: this.options,
-      params,
+      queryKey,
       state: this.state,
-      queryFn,
-    }
+      fetchFn,
+    };
 
     if (this.options.behavior?.onFetch) {
-      this.options.behavior?.onFetch(context)
+      this.options.behavior?.onFetch(context);
     }
 
     // Set to fetching state if not already in it
@@ -368,84 +384,110 @@ export class Query<TData = unknown, TError = unknown, TQueryFnData = TData> {
       !this.state.isFetching ||
       this.state.fetchMeta !== context.fetchOptions?.meta
     ) {
-      this.dispatch({ type: 'fetch', meta: context.fetchOptions?.meta })
+      this.dispatch({ type: 'fetch', meta: context.fetchOptions?.meta });
     }
 
     // Try to fetch the data
     this.retryer = new Retryer({
-      fn: context.queryFn,
+      fn: context.fetchFn,
       onFail: () => {
-        this.dispatch({ type: 'failed' })
+        this.dispatch({ type: 'failed' });
       },
       onPause: () => {
-        this.dispatch({ type: 'pause' })
+        this.dispatch({ type: 'pause' });
       },
       onContinue: () => {
-        this.dispatch({ type: 'continue' })
+        this.dispatch({ type: 'continue' });
       },
       retry: context.options.retry,
       retryDelay: context.options.retryDelay,
-    })
+    });
 
     this.promise = this.retryer.promise
-      .then(data => this.setData(data as TData))
-      .catch(error => {
+      .then((data) => this.setData(data as TData))
+      .catch((error) => {
         // Set error state if needed
         if (!(isCancelledError(error) && error.silent)) {
           this.dispatch({
             type: 'error',
             error,
-          })
+          });
         }
 
-        // Log error
         if (!isCancelledError(error)) {
-          getLogger().error(error)
+          // Notify cache callback
+          if (this.cache.config.onError) {
+            this.cache.config.onError(error, this as Query);
+          }
+
+          // Log error
+          getLogger().error(error);
+        }
+
+        // Remove query after fetching if cache time is 0
+        if (this.cacheTime === 0) {
+          this.optionalRemove();
         }
 
         // Propagate error
-        throw error
+        throw error;
       })
+      .then((data) => {
+        // Remove query after fetching if cache time is 0
+        if (this.cacheTime === 0) {
+          this.optionalRemove();
+        }
 
-    return this.promise
+        return data;
+      });
+
+    return this.promise;
   }
 
   private dispatch(action: Action<TData, TError>): void {
-    this.state = this.reducer(this.state, action)
+    this.state = this.reducer(this.state, action);
 
     notifyManager.batch(() => {
-      this.observers.forEach(observer => {
-        observer.onQueryUpdate(action)
-      })
+      this.observers.forEach((observer) => {
+        observer.onQueryUpdate(action);
+      });
 
-      this.cache.notify(this)
-    })
+      this.cache.notify(this);
+    });
   }
 
   protected getDefaultState(
-    options: QueryOptions<TData, TError, TQueryFnData>
+    options: QueryOptions<TQueryFnData, TError, TData>
   ): QueryState<TData, TError> {
     const data =
       typeof options.initialData === 'function'
         ? (options.initialData as InitialDataFunction<TData>)()
-        : options.initialData
+        : options.initialData;
 
-    const hasData = typeof data !== 'undefined'
+    const hasInitialData = typeof options.initialData !== 'undefined';
+
+    const initialDataUpdatedAt =
+      hasInitialData &&
+      (typeof options.initialDataUpdatedAt === 'function'
+        ? (options.initialDataUpdatedAt as () => number | undefined)()
+        : options.initialDataUpdatedAt);
+
+    const hasData = typeof data !== 'undefined';
 
     return {
       data,
       dataUpdateCount: 0,
-      dataUpdatedAt: hasData ? Date.now() : 0,
+      dataUpdatedAt: hasData ? initialDataUpdatedAt || Date.now() : 0,
       error: null,
       errorUpdateCount: 0,
       errorUpdatedAt: 0,
       fetchFailureCount: 0,
-      fetchMeta: undefined,
+      fetchMeta: null,
       isFetching: false,
       isInvalidated: false,
       isPaused: false,
       status: hasData ? 'success' : 'idle',
-    }
+    };
   }
 
   protected reducer(
@@ -457,50 +499,60 @@ export class Query<TData = unknown, TError = unknown, TQueryFnData = TData> {
         return {
           ...state,
           fetchFailureCount: state.fetchFailureCount + 1,
-        }
+        };
       case 'pause':
         return {
           ...state,
           isPaused: true,
-        }
+        };
       case 'continue':
         return {
           ...state,
           isPaused: false,
-        }
+        };
       case 'fetch':
         return {
           ...state,
           fetchFailureCount: 0,
-          fetchMeta: action.meta,
+          fetchMeta: action.meta ?? null,
           isFetching: true,
           isPaused: false,
-          status: state.status === 'idle' ? 'loading' : state.status,
-        }
+          status: !state.dataUpdatedAt ? 'loading' : state.status,
+        };
       case 'success':
         return {
           ...state,
           data: action.data,
           dataUpdateCount: state.dataUpdateCount + 1,
-          dataUpdatedAt: action.updatedAt ?? Date.now(),
+          dataUpdatedAt: action.dataUpdatedAt ?? Date.now(),
           error: null,
           fetchFailureCount: 0,
           isFetching: false,
           isInvalidated: false,
           isPaused: false,
           status: 'success',
-        }
+        };
       case 'error':
-        const error = action.error as unknown
+        const error = action.error as unknown;
 
         if (isCancelledError(error) && error.revert) {
+          let previousStatus: QueryStatus;
+
+          if (!state.dataUpdatedAt && !state.errorUpdatedAt) {
+            previousStatus = 'idle';
+          } else if (state.dataUpdatedAt > state.errorUpdatedAt) {
+            previousStatus = 'success';
+          } else {
+            previousStatus = 'error';
+          }
+
           return {
             ...state,
             fetchFailureCount: 0,
             isFetching: false,
             isPaused: false,
-            status: state.status === 'loading' ? 'idle' : state.status,
-          }
+            status: previousStatus,
+          };
         }
 
         return {
@@ -512,19 +564,19 @@ export class Query<TData = unknown, TError = unknown, TQueryFnData = TData> {
           isFetching: false,
           isPaused: false,
           status: 'error',
-        }
+        };
       case 'invalidate':
         return {
           ...state,
           isInvalidated: true,
-        }
+        };
       case 'setState':
         return {
           ...state,
           ...action.state,
-        }
+        };
       default:
-        return state
+        return state;
     }
   }
 }
