@@ -3,43 +3,49 @@ import {
   getQueryKeyHashFn,
   matchQuery,
   parseFilterArgs,
-} from './utils'
-import { Query, QueryState } from './query'
-import type { QueryKey, QueryOptions } from './types'
-import { notifyManager } from './notifyManager'
-import type { QueryClient } from './queryClient'
-import { Subscribable } from './subscribable'
+} from './utils';
+import { Query, QueryState } from './query';
+import type { QueryKey, QueryOptions } from './types';
+import { notifyManager } from './notifyManager';
+import type { QueryClient } from './queryClient';
+import { Subscribable } from './subscribable';
 
 // TYPES
 
-interface QueryHashMap {
-  [hash: string]: Query<any, any>
+interface QueryCacheConfig {
+  onError?: (error: unknown, query: Query<unknown, unknown, unknown>) => void;
 }
 
-type QueryCacheListener = (query?: Query) => void
+interface QueryHashMap {
+  [hash: string]: Query<any, any>;
+}
+
+type QueryCacheListener = (query?: Query) => void;
 
 // CLASS
 
 export class QueryCache extends Subscribable<QueryCacheListener> {
-  private queries: Query<any, any>[]
-  private queriesMap: QueryHashMap
+  config: QueryCacheConfig;
 
-  constructor() {
-    super()
+  private queries: Query<any, any>[];
+  private queriesMap: QueryHashMap;
 
-    this.queries = []
-    this.queriesMap = {}
+  constructor(config?: QueryCacheConfig) {
+    super();
+    this.config = config || {};
+    this.queries = [];
+    this.queriesMap = {};
   }
 
-  build<TData, TError, TQueryFnData>(
+  build<TQueryFnData, TError, TData>(
     client: QueryClient,
-    options: QueryOptions<TData, TError, TQueryFnData>,
+    options: QueryOptions<TQueryFnData, TError, TData>,
     state?: QueryState<TData, TError>
-  ): Query<TData, TError, TQueryFnData> {
-    const hashFn = getQueryKeyHashFn(options)
-    const queryKey = options.queryKey!
-    const queryHash = options.queryHash ?? hashFn(queryKey)
-    let query = this.get<TData, TError, TQueryFnData>(queryHash)
+  ): Query<TQueryFnData, TError, TData> {
+    const hashFn = getQueryKeyHashFn(options);
+    const queryKey = options.queryKey!;
+    const queryHash = options.queryHash ?? hashFn(queryKey);
+    let query = this.get<TQueryFnData, TError, TData>(queryHash);
 
     if (!query) {
       query = new Query({
@@ -49,87 +55,94 @@ export class QueryCache extends Subscribable<QueryCacheListener> {
         options: client.defaultQueryOptions(options),
         state,
         defaultOptions: client.getQueryDefaults(queryKey),
-      })
-      this.add(query)
+      });
+      this.add(query);
     }
 
-    return query
+    return query;
   }
 
   add(query: Query<any, any>): void {
     if (!this.queriesMap[query.queryHash]) {
-      this.queriesMap[query.queryHash] = query
-      this.queries.push(query)
-      this.notify(query)
+      this.queriesMap[query.queryHash] = query;
+      this.queries.push(query);
+      this.notify(query);
     }
   }
 
   remove(query: Query<any, any>): void {
-    if (this.queriesMap[query.queryHash]) {
-      query.destroy()
-      delete this.queriesMap[query.queryHash]
-      this.queries = this.queries.filter(x => x !== query)
-      this.notify(query)
+    const queryInMap = this.queriesMap[query.queryHash];
+
+    if (queryInMap) {
+      query.destroy();
+
+      this.queries = this.queries.filter((x) => x !== query);
+
+      if (queryInMap === query) {
+        delete this.queriesMap[query.queryHash];
+      }
+
+      this.notify(query);
     }
   }
 
   clear(): void {
     notifyManager.batch(() => {
-      this.queries.forEach(query => {
-        this.remove(query)
-      })
-    })
+      this.queries.forEach((query) => {
+        this.remove(query);
+      });
+    });
   }
 
-  get<TData = unknown, TError = unknown, TQueryFnData = TData>(
+  get<TQueryFnData = unknown, TError = unknown, TData = TQueryFnData>(
     queryHash: string
-  ): Query<TData, TError, TQueryFnData> | undefined {
-    return this.queriesMap[queryHash]
+  ): Query<TQueryFnData, TError, TData> | undefined {
+    return this.queriesMap[queryHash];
   }
 
   getAll(): Query[] {
-    return this.queries
+    return this.queries;
   }
 
-  find<TData = unknown, TError = unknown, TQueryFnData = TData>(
+  find<TQueryFnData = unknown, TError = unknown, TData = TQueryFnData>(
     arg1: QueryKey,
     arg2?: QueryFilters
-  ): Query<TData, TError, TQueryFnData> | undefined {
-    const [filters] = parseFilterArgs(arg1, arg2)
-    return this.queries.find(query => matchQuery(filters, query))
+  ): Query<TQueryFnData, TError, TData> | undefined {
+    const [filters] = parseFilterArgs(arg1, arg2);
+    return this.queries.find((query) => matchQuery(filters, query));
   }
 
-  findAll(queryKey?: QueryKey, filters?: QueryFilters): Query[]
-  findAll(filters?: QueryFilters): Query[]
-  findAll(arg1?: QueryKey | QueryFilters, arg2?: QueryFilters): Query[]
+  findAll(queryKey?: QueryKey, filters?: QueryFilters): Query[];
+  findAll(filters?: QueryFilters): Query[];
+  findAll(arg1?: QueryKey | QueryFilters, arg2?: QueryFilters): Query[];
   findAll(arg1?: QueryKey | QueryFilters, arg2?: QueryFilters): Query[] {
-    const [filters] = parseFilterArgs(arg1, arg2)
+    const [filters] = parseFilterArgs(arg1, arg2);
     return filters
-      ? this.queries.filter(query => matchQuery(filters, query))
-      : this.queries
+      ? this.queries.filter((query) => matchQuery(filters, query))
+      : this.queries;
   }
 
   notify(query?: Query<any, any>) {
     notifyManager.batch(() => {
-      this.listeners.forEach(listener => {
-        listener(query)
-      })
-    })
+      this.listeners.forEach((listener) => {
+        listener(query);
+      });
+    });
   }
 
   onFocus(): void {
     notifyManager.batch(() => {
-      this.queries.forEach(query => {
-        query.onFocus()
-      })
-    })
+      this.queries.forEach((query) => {
+        query.onFocus();
+      });
+    });
   }
 
   onOnline(): void {
     notifyManager.batch(() => {
-      this.queries.forEach(query => {
-        query.onOnline()
-      })
-    })
+      this.queries.forEach((query) => {
+        query.onOnline();
+      });
+    });
   }
 }
