@@ -1,4 +1,5 @@
-import { RetryValue, RetryDelayValue } from './query/core/retryer';
+import { Observable } from 'rxjs';
+import { RetryDelayValue, RetryValue } from './query/core/retryer';
 import {
   InfiniteQueryObserverOptions,
   InfiniteQueryObserverResult,
@@ -8,31 +9,68 @@ import {
   QueryObserverResult,
 } from './query/core/types';
 
-export interface UseBaseQueryOptions<TData = unknown, TError = unknown, TQueryFnData = TData, TQueryData = TQueryFnData>
-  extends QueryObserverOptions<TData, TError, TQueryFnData, TQueryData> {}
+export type QueryFunctionWithObservable<T = unknown> = (
+  ...args: any[]
+) => T | Promise<T> | Observable<T>;
 
-export interface UseQueryOptions<TData = unknown, TError = unknown, TQueryFnData = TData>
-  extends UseBaseQueryOptions<TData, TError, TQueryFnData> {}
-
-export interface UseInfiniteQueryOptions<
+export interface UseBaseQueryOptions<
   TData = unknown,
   TError = unknown,
   TQueryFnData = TData,
   TQueryData = TQueryFnData
-> extends InfiniteQueryObserverOptions<TData, TError, TQueryFnData, TQueryData> {}
+> extends Omit<
+    QueryObserverOptions<TData, TError, TQueryFnData, TQueryData>,
+    'queryFn'
+  > {
+  queryFn?: QueryFunctionWithObservable<TQueryFnData>;
+}
 
-export interface UseBaseQueryResult<TData = unknown, TError = unknown> extends QueryObserverResult<TData, TError> {}
+export type UseQueryOptions<
+  TData = unknown,
+  TError = unknown,
+  TQueryFnData = TData
+> = UseBaseQueryOptions<TData, TError, TQueryFnData>;
 
-export interface UseQueryResult<TData = unknown, TError = unknown> extends UseBaseQueryResult<TData, TError> {}
+export type UseInfiniteQueryOptions<
+  TData = unknown,
+  TError = unknown,
+  TQueryFnData = TData,
+  TQueryData = TQueryFnData
+> = InfiniteQueryObserverOptions<TData, TError, TQueryFnData, TQueryData>;
 
-export interface UseInfiniteQueryResult<TData = unknown, TError = unknown>
-  extends InfiniteQueryObserverResult<TData, TError> {}
+export type UseBaseQueryResult<
+  TData = unknown,
+  TError = unknown
+> = QueryObserverResult<TData, TError>;
 
-export interface UseMutationOptions<TData = unknown, TError = unknown, TVariables = void, TContext = unknown> {
+export type UseQueryResult<
+  TData = unknown,
+  TError = unknown
+> = UseBaseQueryResult<TData, TError>;
+
+export type UseInfiniteQueryResult<
+  TData = unknown,
+  TError = unknown
+> = InfiniteQueryObserverResult<TData, TError>;
+
+export interface UseMutationOptions<
+  TData = unknown,
+  TError = unknown,
+  TVariables = void,
+  TContext = unknown
+> {
   mutationKey?: string | unknown[];
   onMutate?: (variables: TVariables) => Promise<TContext> | TContext;
-  onSuccess?: (data: TData, variables: TVariables, context: TContext | undefined) => Promise<void> | void;
-  onError?: (error: TError, variables: TVariables, context: TContext | undefined) => Promise<void> | void;
+  onSuccess?: (
+    data: TData,
+    variables: TVariables,
+    context: TContext | undefined
+  ) => Promise<void> | void;
+  onError?: (
+    error: TError,
+    variables: TVariables,
+    context: TContext | undefined
+  ) => Promise<void> | void;
   onSettled?: (
     data: TData | undefined,
     error: TError | null,
@@ -44,17 +82,32 @@ export interface UseMutationOptions<TData = unknown, TError = unknown, TVariable
   useErrorBoundary?: boolean;
 }
 
-export type UseMutateFunction<TData = unknown, TError = unknown, TVariables = void, TContext = unknown> = (
+export type UseMutateFunction<
+  TData = unknown,
+  TError = unknown,
+  TVariables = void,
+  TContext = unknown
+> = (
   variables: TVariables,
   options?: MutateOptions<TData, TError, TVariables, TContext>
 ) => void;
 
-export type UseMutateAsyncFunction<TData = unknown, TError = unknown, TVariables = void, TContext = unknown> = (
+export type UseMutateAsyncFunction<
+  TData = unknown,
+  TError = unknown,
+  TVariables = void,
+  TContext = unknown
+> = (
   variables: TVariables,
   options?: MutateOptions<TData, TError, TVariables, TContext>
 ) => Promise<TData>;
 
-export interface UseMutationResult<TData = unknown, TError = unknown, TVariables = unknown, TContext = unknown> {
+export interface UseMutationResult<
+  TData = unknown,
+  TError = unknown,
+  TVariables = unknown,
+  TContext = unknown
+> {
   context: TContext | undefined;
   data: TData | undefined;
   error: TError | null;
