@@ -2,9 +2,9 @@
  * Public API Surface of angular-query
  */
 
-export * from './lib/use-query/use-query.service';
-export * from './lib/use-mutation/use-mutation.service';
-export * from './lib/use-base-query/use-base-query.service';
 export * from './lib/angular-query.module';
-export * from './lib/types';
 export * from './lib/query';
+export * from './lib/types';
+export * from './lib/use-base-query/use-base-query.service';
+export * from './lib/use-mutation/use-mutation.service';
+export * from './lib/use-query/use-query.service';
